Updating role to new purpose
Support for internal and external roles is no longer needed as all our roles will be open sourced. This drastically reduces the complexity of this role (for the better).

It now serves two purposes, get roles from Ansible Galaxy (from a since roles file) and copy the users public key into the project (as it did before).

Everything else can then be configured as before. Outside of this role, a new prelude playbook has been developed to remove the need for Vagrant-Triggers.

This role has also be brought up to date with the latest conventions for documentation and project structure. As there are breaking changes, this release bumps the major version.