Fixing Ansible Galaxy task
This needs to be run from the root of the project so the ansible.cfg file, with a non-default roles directory, is picked up. Otherwise roles are installed to the default system wide location, but not used when ansible-playbook is run as the ansible.cfg is then used, so its looking in the custom roles directory where the roles won't be.